Potential fix for SDL crash on boot.
The root cause of the core crash could be related to
lazy initialization of the HMI/Mobile API classes.
These classes are initializated on demand when some
of threads are trying to get the object of that class
first time. Most likely at SDL start could be a situation
when HMI/Mobile API classes are constructing in one thread
and at that moment another thread is trying to get access to
the inner data of not-yet-fully-constructed object and this
causes a core crash with ~ISchemaItem() calls in stack trace.